Refactor worker channel sync ownership
