+++ src/prefect/workers/_worker_channel/__init__.py
@@ -1,0 +1,37 @@
+from __future__ import annotations
+
+from prefect.workers._worker_channel._protocol import WorkerChannelProtocolHandler
+from prefect.workers._worker_channel._state import (
+    WorkerChannel,
+    WorkerChannelConnection,
+    WorkerChannelError,
+    WorkerChannelFallbackState,
+    WorkerChannelRetryableError,
+    WorkerChannelState,
+    WorkerChannelStatus,
+    WorkerChannelTerminalError,
+)
+from prefect.workers._worker_channel._sync import WorkPoolWorkerChannel
+from prefect.workers._worker_channel._transport import (
+    WORKER_CHANNEL_SETUP_TIMEOUT_SECONDS,
+    ConnectFactory,
+    WorkerChannelTransport,
+    build_worker_channel_url,
+)
+
+__all__ = [
+    "WORKER_CHANNEL_SETUP_TIMEOUT_SECONDS",
+    "ConnectFactory",
+    "WorkerChannel",
+    "WorkerChannelConnection",
+    "WorkerChannelError",
+    "WorkerChannelFallbackState",
+    "WorkerChannelProtocolHandler",
+    "WorkerChannelRetryableError",
+    "WorkerChannelState",
+    "WorkerChannelStatus",
+    "WorkerChannelTerminalError",
+    "WorkerChannelTransport",
+    "WorkPoolWorkerChannel",
+    "build_worker_channel_url",
+]
